restructure headerfiles of SANSData
