+++ src/OutputData/SANSData/NuMagSANSlib_SANSData.h
@@ -28,124 +28,16 @@
 #include <filesystem>
 #include <iomanip>
 
+
+#include "NuMagSANSlib_SANSDataStruct.h"
+#include "NuMagSANSlib_SANSData2csv.h"
+
 using namespace std;
 
-struct ScatteringData {
 
-    unsigned int* N_q;     // number of q-values
-    unsigned int* N_theta; // number of theta-values
-    unsigned int* N_r;     // number of r-values
-    unsigned int* N_alpha; // number of alpha-values
 
-    float* Polarization; // polarization vector [Px, Py, Pz]
 
-    float* q_min; // minimum q-value
-    float* q_max; // maximum q-value
-    float* r_max; // maximum r-value
 
-    float* dq;     // step-size of q
-    float* dtheta; // step-size of theta
-    float* dr;     // step-size of r
-    float* dalpha; // step-size of alpha
-
-    float* qy_2D;    // 2D qy-scattering vector
-    float* qz_2D;    // 2D qz-scattering vector
-    float* q_2D;     // 2D q-scattering vector
-    float* theta_2D; // 2D theta-scattering vector
-
-    float* ry_2D;    // 2D ry-correlation space coordinate
-    float* rz_2D;    // 2D rz-correlation space coordinate
-    float* r_2D;     // 2D r-correlation space coordinate
-    float* alpha_2D; // 2D angle-correlation space coordinate
-
-    float* q_1D; // 1D scattering vector
-    float* r_1D; // 1D correlation space coordinate
-
-    float* Gxx_real; // real-part xx-magnetic Fourier correlation function
-    float* Gyy_real; // real-part yy-magnetic Fourier correlation function
-    float* Gzz_real; // real part zz-magnetic Fourier correlation function
-    float* Gxy_real; // real part xy-magnetic Fourier correlation function
-    float* Gyx_real; // real part yx-magnetic Fourier correlation function
-    float* Gxz_real; // real part xz-magnetic Fourier correlation function
-    float* Gzx_real; // real part zx-magnetic Fourier correlation function
-    float* Gyz_real; // real part yz-magnetic Fourier correlation function
-    float* Gzy_real; // real part zy-magnetic Fourier correlation function
-
-    float* Gxx_imag; // real-part xx-magnetic Fourier correlation function
-    float* Gyy_imag; // real-part yy-magnetic Fourier correlation function
-    float* Gzz_imag; // real-part zz-magnetic Fourier correlation function
-    float* Gxy_imag; // real-part xy-magnetic Fourier correlation function
-    float* Gyx_imag; // real-part yx-magnetic Fourier correlation function
-    float* Gxz_imag; // real-part xz-magnetic Fourier correlation function
-    float* Gzx_imag; // real-part zx-magnetic Fourier correlation function
-    float* Gyz_imag; // real-part yz-magnetic Fourier correlation function
-    float* Gzy_imag; // real-part zy-magnetic Fourier correlation function
-
-    float* S_Nuc_2D_unpolarized;      // nuclear SANS cross section
-    float* S_Mag_2D_unpolarized;      // unpolarized magnetic SANS cross section
-    float* S_Mag_2D_polarized;        // polarized magnetic SANS cross section
-    float* S_NucMag_2D;               // nuclear-magnetic interference SANS cross section
-    float* S_Mag_2D_spin_flip;        // spin-flip magnetic SANS cross section
-    float* S_Mag_2D_chiral;           // chiral magnetic SANS cross section
-    float* S_Mag_2D_spin_flip_pm;     // pm-spin-flip magnetic SANS cross section
-    float* S_Mag_2D_spin_flip_mp;     // mp-spin-flip magnetic SANS cross section
-    float* S_Mag_2D_non_spin_flip_pp; // pp-non-spin-flip magnetic SANS cross section
-    float* S_Mag_2D_non_spin_flip_mm; // mm-non-spin-flip magnetic SANS cross section
-    float* S_Mag_2D_sanspol_p;        // p-sanspol magnetic SANS cross section
-    float* S_Mag_2D_sanspol_m;        // m-sanspol magnetic SANS cross section
-
-    float* Corr_Nuc_2D_unpolarized;
-    float* Corr_Mag_2D_unpolarized;
-    float* Corr_Mag_2D_polarized;
-    float* Corr_NucMag_2D;
-    float* Corr_Mag_2D_spin_flip;
-    float* Corr_Mag_2D_chiral;
-    float* Corr_Mag_2D_spin_flip_pm;
-    float* Corr_Mag_2D_spin_flip_mp;
-    float* Corr_Mag_2D_non_spin_flip_pp;
-    float* Corr_Mag_2D_non_spin_flip_mm;
-    float* Corr_Mag_2D_sanspol_p;
-    float* Corr_Mag_2D_sanspol_m;
-
-    float* S_Nuc_1D_unpolarized;
-    float* S_Mag_1D_unpolarized;
-    float* S_Mag_1D_polarized;
-    float* S_NucMag_1D;
-    float* S_Mag_1D_chiral;
-    float* S_Mag_1D_spin_flip;
-    float* S_Mag_1D_spin_flip_pm;
-    float* S_Mag_1D_spin_flip_mp;
-    float* S_Mag_1D_non_spin_flip_pp;
-    float* S_Mag_1D_non_spin_flip_mm;
-    float* S_Mag_1D_sanspol_p;
-    float* S_Mag_1D_sanspol_m;
-
-    float* p_Nuc_unpolarized;
-    float* p_Mag_unpolarized;
-    float* p_Mag_polarized;
-    float* p_NucMag;
-    float* p_Mag_chiral;
-    float* p_Mag_spin_flip;
-    float* p_Mag_spin_flip_pm;
-    float* p_Mag_spin_flip_mp;
-    float* p_Mag_non_spin_flip_pp;
-    float* p_Mag_non_spin_flip_mm;
-    float* p_Mag_sanspol_p;
-    float* p_Mag_sanspol_m;
-
-    float* c_Nuc_unpolarized;
-    float* c_Mag_unpolarized;
-    float* c_Mag_polarized;
-    float* c_NucMag;
-    float* c_Mag_chiral;
-    float* c_Mag_spin_flip;
-    float* c_Mag_spin_flip_pm;
-    float* c_Mag_spin_flip_mp;
-    float* c_Mag_non_spin_flip_pp;
-    float* c_Mag_non_spin_flip_mm;
-    float* c_Mag_sanspol_p;
-    float* c_Mag_sanspol_m;
-};
 
 // #################################################################################################################################################
 struct ScalingFactors {
@@ -1098,307 +990,6 @@ void copyGPU2RAM_ScatteringData(ScatteringData* SANSData, ScatteringData* SANSDa
     cudaMemcpy(SANSData->c_Mag_sanspol_m, SANSData_gpu->c_Mag_sanspol_m, (*SANSData->N_r) * sizeof(float),
                cudaMemcpyDeviceToHost);
 }
-
-// #################################################################################################
-// CSV Writer Backend ###################################################################################
-// #################################################################################################
-
-// -------------------------------------------------------------------------------------------------
-struct Column {
-    std::string name;
-    const float* data;
-    Column(const std::string& n, const float* d) : name(n), data(d) {}
-};
-// -------------------------------------------------------------------------------------------------
-
-void writeCSV(const std::string& filename, unsigned long length, const std::vector<Column>& columns) {
-    std::ofstream fout(filename);
-    if (!fout.is_open()) {
-        throw std::runtime_error("Could not open file: " + filename);
-    }
-
-    // Header
-    for (size_t i = 0; i < columns.size(); ++i) {
-        fout << columns[i].name;
-        if (i < columns.size() - 1)
-            fout << ",";
-    }
-    fout << "\n";
-
-    // Data
-    for (unsigned long n = 0; n < length; ++n) {
-        for (size_t i = 0; i < columns.size(); ++i) {
-            fout << columns[i].data[n];
-            if (i < columns.size() - 1)
-                fout << ",";
-        }
-        fout << "\n";
-    }
-
-    fout.close();
-}
-
-std::vector<Column> build_SANS2D_columns(InputFileData* InputData, ScatteringData* SANSData) {
-    std::vector<Column> columns;
-
-    // Coordinates (immer)
-    columns.emplace_back("qz", SANSData->qz_2D);
-    columns.emplace_back("qy", SANSData->qy_2D);
-    columns.emplace_back("q", SANSData->q_2D);
-    columns.emplace_back("theta", SANSData->theta_2D);
-
-    // Fourier matrix
-    if (InputData->output_fourier_correlation_matrix_flag) {
-        columns.emplace_back("Gxx_real", SANSData->Gxx_real);
-        columns.emplace_back("Gxx_imag", SANSData->Gxx_imag);
-        columns.emplace_back("Gyy_real", SANSData->Gyy_real);
-        columns.emplace_back("Gyy_imag", SANSData->Gyy_imag);
-        columns.emplace_back("Gzz_real", SANSData->Gzz_real);
-        columns.emplace_back("Gzz_imag", SANSData->Gzz_imag);
-        columns.emplace_back("Gxy_real", SANSData->Gxy_real);
-        columns.emplace_back("Gxy_imag", SANSData->Gxy_imag);
-        columns.emplace_back("Gyx_real", SANSData->Gyx_real);
-        columns.emplace_back("Gyx_imag", SANSData->Gyx_imag);
-        columns.emplace_back("Gxz_real", SANSData->Gxz_real);
-        columns.emplace_back("Gxz_imag", SANSData->Gxz_imag);
-        columns.emplace_back("Gzx_real", SANSData->Gzx_real);
-        columns.emplace_back("Gzx_imag", SANSData->Gzx_imag);
-        columns.emplace_back("Gyz_real", SANSData->Gyz_real);
-        columns.emplace_back("Gyz_imag", SANSData->Gyz_imag);
-        columns.emplace_back("Gzy_real", SANSData->Gzy_real);
-        columns.emplace_back("Gzy_imag", SANSData->Gzy_imag);
-    }
-
-    auto& f = InputData->OutFlags.SANS2D;
-
-    if (f.Nuclear)
-        columns.emplace_back("S_N", SANSData->S_Nuc_2D_unpolarized);
-    if (f.Unpolarized)
-        columns.emplace_back("S_M", SANSData->S_Mag_2D_unpolarized);
-    if (f.NuclearMagnetic)
-        columns.emplace_back("S_NM", SANSData->S_NucMag_2D);
-    if (f.Polarized)
-        columns.emplace_back("S_P", SANSData->S_Mag_2D_polarized);
-    if (f.Chiral)
-        columns.emplace_back("S_chi", SANSData->S_Mag_2D_chiral);
-    if (f.SpinFlip)
-        columns.emplace_back("S_sf", SANSData->S_Mag_2D_spin_flip);
-    if (f.PM_SpinFlip)
-        columns.emplace_back("S_pm", SANSData->S_Mag_2D_spin_flip_pm);
-    if (f.MP_SpinFlip)
-        columns.emplace_back("S_mp", SANSData->S_Mag_2D_spin_flip_mp);
-    if (f.PP_NonSpinFlip)
-        columns.emplace_back("S_pp", SANSData->S_Mag_2D_non_spin_flip_pp);
-    if (f.MM_NonSpinFlip)
-        columns.emplace_back("S_mm", SANSData->S_Mag_2D_non_spin_flip_mm);
-    if (f.P_SANSPOL)
-        columns.emplace_back("S_p", SANSData->S_Mag_2D_sanspol_p);
-    if (f.M_SANSPOL)
-        columns.emplace_back("S_m", SANSData->S_Mag_2D_sanspol_m);
-
-    return columns;
-}
-
-std::vector<Column> build_SANS1D_columns(InputFileData* InputData, ScatteringData* SANSData) {
-    std::vector<Column> columns;
-
-    columns.emplace_back("q", SANSData->q_1D);
-
-    auto& f = InputData->OutFlags.SANS1D;
-
-    if (f.Nuclear)
-        columns.emplace_back("I_N", SANSData->S_Nuc_1D_unpolarized);
-    if (f.Unpolarized)
-        columns.emplace_back("I_M", SANSData->S_Mag_1D_unpolarized);
-    if (f.NuclearMagnetic)
-        columns.emplace_back("I_NM", SANSData->S_NucMag_1D);
-    if (f.Polarized)
-        columns.emplace_back("I_P", SANSData->S_Mag_1D_polarized);
-    if (f.Chiral)
-        columns.emplace_back("I_chi", SANSData->S_Mag_1D_chiral);
-    if (f.SpinFlip)
-        columns.emplace_back("I_sf", SANSData->S_Mag_1D_spin_flip);
-    if (f.PM_SpinFlip)
-        columns.emplace_back("I_pm", SANSData->S_Mag_1D_spin_flip_pm);
-    if (f.MP_SpinFlip)
-        columns.emplace_back("I_mp", SANSData->S_Mag_1D_spin_flip_mp);
-    if (f.PP_NonSpinFlip)
-        columns.emplace_back("I_pp", SANSData->S_Mag_1D_non_spin_flip_pp);
-    if (f.MM_NonSpinFlip)
-        columns.emplace_back("I_mm", SANSData->S_Mag_1D_non_spin_flip_mm);
-    if (f.P_SANSPOL)
-        columns.emplace_back("I_p", SANSData->S_Mag_1D_sanspol_p);
-    if (f.M_SANSPOL)
-        columns.emplace_back("I_m", SANSData->S_Mag_1D_sanspol_m);
-
-    return columns;
-}
-
-std::vector<Column> build_Corr1D_columns(InputFileData* InputData, ScatteringData* SANSData) {
-    std::vector<Column> columns;
-
-    // r always first
-    columns.emplace_back("r", SANSData->r_1D);
-
-    auto& p = InputData->OutFlags.PairDist1D;
-    auto& c = InputData->OutFlags.Corr1D;
-
-    // ---------- Pair Distribution p_*
-    if (p.Nuclear)
-        columns.emplace_back("p_N", SANSData->p_Nuc_unpolarized);
-    if (p.Unpolarized)
-        columns.emplace_back("p_M", SANSData->p_Mag_unpolarized);
-    if (p.NuclearMagnetic)
-        columns.emplace_back("p_NM", SANSData->p_NucMag);
-    if (p.Polarized)
-        columns.emplace_back("p_P", SANSData->p_Mag_polarized);
-    if (p.Chiral)
-        columns.emplace_back("p_chi", SANSData->p_Mag_chiral);
-    if (p.SpinFlip)
-        columns.emplace_back("p_sf", SANSData->p_Mag_spin_flip);
-    if (p.PM_SpinFlip)
-        columns.emplace_back("p_pm", SANSData->p_Mag_spin_flip_pm);
-    if (p.MP_SpinFlip)
-        columns.emplace_back("p_mp", SANSData->p_Mag_spin_flip_mp);
-    if (p.PP_NonSpinFlip)
-        columns.emplace_back("p_pp", SANSData->p_Mag_non_spin_flip_pp);
-    if (p.MM_NonSpinFlip)
-        columns.emplace_back("p_mm", SANSData->p_Mag_non_spin_flip_mm);
-    if (p.P_SANSPOL)
-        columns.emplace_back("p_p", SANSData->p_Mag_sanspol_p);
-    if (p.M_SANSPOL)
-        columns.emplace_back("p_m", SANSData->p_Mag_sanspol_m);
-
-    // ---------- Correlation c_*
-    if (c.Nuclear)
-        columns.emplace_back("c_N", SANSData->c_Nuc_unpolarized);
-    if (c.Unpolarized)
-        columns.emplace_back("c_M", SANSData->c_Mag_unpolarized);
-    if (c.NuclearMagnetic)
-        columns.emplace_back("c_NM", SANSData->c_NucMag);
-    if (c.Polarized)
-        columns.emplace_back("c_P", SANSData->c_Mag_polarized);
-    if (c.Chiral)
-        columns.emplace_back("c_chi", SANSData->c_Mag_chiral);
-    if (c.SpinFlip)
-        columns.emplace_back("c_sf", SANSData->c_Mag_spin_flip);
-    if (c.PM_SpinFlip)
-        columns.emplace_back("c_pm", SANSData->c_Mag_spin_flip_pm);
-    if (c.MP_SpinFlip)
-        columns.emplace_back("c_mp", SANSData->c_Mag_spin_flip_mp);
-    if (c.PP_NonSpinFlip)
-        columns.emplace_back("c_pp", SANSData->c_Mag_non_spin_flip_pp);
-    if (c.MM_NonSpinFlip)
-        columns.emplace_back("c_mm", SANSData->c_Mag_non_spin_flip_mm);
-    if (c.P_SANSPOL)
-        columns.emplace_back("c_p", SANSData->c_Mag_sanspol_p);
-    if (c.M_SANSPOL)
-        columns.emplace_back("c_m", SANSData->c_Mag_sanspol_m);
-
-    return columns;
-}
-
-std::vector<Column> build_Corr2D_columns(InputFileData* InputData, ScatteringData* SANSData) {
-    std::vector<Column> columns;
-
-    // ---- Coordinates (immer)
-    columns.emplace_back("rz", SANSData->rz_2D);
-    columns.emplace_back("ry", SANSData->ry_2D);
-    columns.emplace_back("r", SANSData->r_2D);
-    columns.emplace_back("alpha", SANSData->alpha_2D);
-
-    auto& f = InputData->OutFlags.Corr2D;
-
-    // ---- Correlation terms
-    if (f.Nuclear)
-        columns.emplace_back("C_N", SANSData->Corr_Nuc_2D_unpolarized);
-    if (f.Unpolarized)
-        columns.emplace_back("C_M", SANSData->Corr_Mag_2D_unpolarized);
-    if (f.NuclearMagnetic)
-        columns.emplace_back("C_NM", SANSData->Corr_NucMag_2D);
-    if (f.Polarized)
-        columns.emplace_back("C_P", SANSData->Corr_Mag_2D_polarized);
-    if (f.Chiral)
-        columns.emplace_back("C_chi", SANSData->Corr_Mag_2D_chiral);
-    if (f.SpinFlip)
-        columns.emplace_back("C_sf", SANSData->Corr_Mag_2D_spin_flip);
-    if (f.PM_SpinFlip)
-        columns.emplace_back("C_pm", SANSData->Corr_Mag_2D_spin_flip_pm);
-    if (f.MP_SpinFlip)
-        columns.emplace_back("C_mp", SANSData->Corr_Mag_2D_spin_flip_mp);
-    if (f.PP_NonSpinFlip)
-        columns.emplace_back("C_pp", SANSData->Corr_Mag_2D_non_spin_flip_pp);
-    if (f.MM_NonSpinFlip)
-        columns.emplace_back("C_mm", SANSData->Corr_Mag_2D_non_spin_flip_mm);
-    if (f.P_SANSPOL)
-        columns.emplace_back("C_p", SANSData->Corr_Mag_2D_sanspol_p);
-    if (f.M_SANSPOL)
-        columns.emplace_back("C_m", SANSData->Corr_Mag_2D_sanspol_m);
-
-    return columns;
-}
-
-void write2CSVtable_ScatteringData(InputFileData* InputData, ScatteringData* SANSData, int MagData_File_Index,
-                                   int StructData_File_Index = 0, int RotData_File_Index = 0) {
-    LogSystem::write("");
-    LogSystem::write("write scattering data to csv-files...");
-
-    std::string target_foldername = InputData->SANSDataFoldername + "/SANS_" + std::to_string(MagData_File_Index) + "/";
-
-    if (StructData_File_Index > 0) {
-        target_foldername += "StructData_" + std::to_string(StructData_File_Index) + "/";
-    }
-
-    if (RotData_File_Index > 0) {
-        target_foldername += "RotData_" + std::to_string(RotData_File_Index) + "/";
-    }
-
-    // mkdir(target_foldername.c_str(), 0777);
-    std::filesystem::create_directories(target_foldername);
-
-    // ------------------- SANS2D -------------------
-    if (InputData->output_fourier_correlation_matrix_flag || any_active(InputData->OutFlags.SANS2D)) {
-        unsigned long L = (*SANSData->N_q) * (*SANSData->N_theta);
-
-        auto columns = build_SANS2D_columns(InputData, SANSData);
-
-        writeCSV(target_foldername + "SANS2D.csv", L, columns);
-
-        LogSystem::write("SANS2D finished...");
-    }
-
-    // ------------------- SANS1D -------------------
-    if (any_active(InputData->OutFlags.SANS1D)) {
-        auto columns = build_SANS1D_columns(InputData, SANSData);
-
-        writeCSV(target_foldername + "SANS1D.csv", (*SANSData->N_q), columns);
-
-        LogSystem::write("SANS1D finished...");
-    }
-
-    // ------------------- Corr1D -------------------
-    if (any_active(InputData->OutFlags.Corr1D) || any_active(InputData->OutFlags.PairDist1D)) {
-        auto columns = build_Corr1D_columns(InputData, SANSData);
-
-        writeCSV(target_foldername + "Corr1D.csv", (*SANSData->N_r), columns);
-
-        LogSystem::write("Corr1D finished...");
-    }
-
-    // ------------------- Corr2D -------------------
-    if (any_active(InputData->OutFlags.Corr2D)) {
-        unsigned long L = (*SANSData->N_r) * (*SANSData->N_alpha);
-
-        auto columns = build_Corr2D_columns(InputData, SANSData);
-
-        writeCSV(target_foldername + "Corr2D.csv", L, columns);
-
-        LogSystem::write("Corr2D finished...");
-    }
-}
-
-// #################################################################################################
-// #################################################################################################
 
 void free_ScatteringData(ScatteringData* SANSData, ScatteringData* SANSData_gpu) {
